add type to the email category
